Add -omit-unchanged-vars option
With this option, if relevance analysis reverts a value all the way to
the preferred (baseline) value, omit it from the testcase entirely.
This makes the testcases shorter and easier to read by hand. In theory
it shouldn't affect their behavior but I'm not sure it won't in
practice.
Also, optimize relevance analysis by checking the full preferred value
first.