Fix longstanding typo in "|" split term case
Simplify in two places in the midst of symbolic address processing,
since with the fancier rules this can sometimes be helpful.
Add another set of simplification rules between casting and masking.
I noticed this was a problem because of an inconsistent number of
test cases generated in phase 1 when using -implied-value-conc.
With this change, the same number 76439 are produced.

Corresponds to upstream commit
f4fe0c0a3a7a057f8f59d8bd00a0f60da8151729