Initial version of Linux 4.16 KVM patch
Based on the Debian Testing 4.16.5 kernel. Two hunks had to be moved
from their locations in the 4.4 patch, but they were just printfs
so it's not too important that they're in the right place.
This version compiles and "insmod"s without problems on my test
machine, but I haven't run tests with it yet.