Zero out uninitialized parts of the KVM sregs state
This seems to be needed to get kvm-run to run on my Linux 4.16-based
system. These fields in the structure seem to have been around for
a while, but maybe older versions didn't check them, or maybe other
system differences made the uninitialized stack contents different.
While I'm at it, add better error messages for two kvm-run failures
that I've been running into, one of which is the symptom of the
problem this main patch fixes.