Remove code to create fake poststate on timeout
This code used to assert that the prestate existed, and then created a
fake poststate. But now because of a new level of temporary files we
use, the prestate does not appear to exist at the location where this
script expects it, and so we couldn't create a poststate even if we
want to. So remove an assertion that was failing on every timeout, and
just finish.