Make a reduced version of the Linux/KVM 4.16 patch for "hlt" only
This removes the behavior of making all exceptions stop KVM's
execution, since I believe that's no longer the behavior we want when
working with modern versions of Fast PokeEMU with their own exception
handling.
Also, disable syslog debugging messages; the one about exceptions in
particular was way to noisy when a test case gets into a state of
raising an exception in a tight loop.