Make the guest kernel Makefile a bit more user friendly
Set the compiler to "gcc-5", to increase the chances of getting a
compatible compiler if available. If gcc-5 is unavailable, the comment
next to this line in the Makefile provides some guidance.

Stop the compilation with an error if the address in the kernel of the
testcase is not 0x219000. This value is hardcoded in other parts of
the system, so things are unlikely to work correctly if it changes.

Automate the copying of "interrupts-mod.c" over "interrupts.c".